refactor(jinja): split templates into files (#117)
* refactor(jinja): use separate files for jinja templates to enable IDE linting and usage of Jinja environments

* include template files in package

* extract generation function and add tests
