+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.1.0.beta3.dev28
+current_version = 0.1.0.beta3.dev29
 commit = True
 tag = True
 parse = (?P<v>v)?(?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)\.?(beta)?(?P<beta>\d+)?\.?(dev)?(?P<dev>\d+)?
