+++ superai/meta_ai/template_contents/entry_script.sh
@@ -1,0 +1,9 @@
+#!/bin/sh
+
+if [ -z "${AWS_LAMBDA_RUNTIME_API}" ]; then
+  # shellcheck disable=SC2068
+  exec /usr/local/bin/aws-lambda-rie /opt/conda/envs/{{env}}/bin/python -m awslambdaric $1
+else
+  # shellcheck disable=SC2068
+  exec /opt/conda/envs/{{env}}/bin/python -m awslambdaric $1
+fi
